+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
 
 - Fix returning `long` update count from update statements (#2778)
 - Fix JFR usage on GraalVM (#2736)
+- Fix SBOM creation
 
 # 3.49.1