+++ RELEASE_NOTES.md
@@ -1,4 +1,4 @@
-# Unreleased
+# 3.49.2
 
 - Fix returning `long` update count from update statements (#2778)
 - Fix JFR usage on GraalVM (#2736)